Updated forecast HTML to 4 days instead of 5
It was too much work to figure out how to fix the formatting of fitting 5 days in the space of 4. I would have built the page differently if I had known I had to build the forecast elements this way. I tried, but I would have had to change too many other things.